requirements.txt
streamlit>=1.33.0
playwright>=1.52.0
pytest>=8.2.0
pyyaml>=6.0.1
pydantic>=2.7.0
reportlab>=4.2.0
requests>=2.32.0
